requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.13.5
bs4==0.0.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
defusedxml==0.7.1
Django==5.2.4
django-cors-headers==4.8.0
django-filter==25.1
django-unused-media==0.2.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.0
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==23.0.0
h11==0.16.0
httpie==3.2.4
idna==3.10
iniconfig==2.1.0
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.6.4
mysqlclient==2.2.7
outcome==1.3.0.post0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
promise==2.3
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
PySocks==1.7.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
rich==14.1.0
selenium==4.34.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
sqlparse==0.5.3
text-unidecode==1.3
tomli==2.2.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
whitenoise==6.9.0
wsproto==1.2.0
